requirements.txt
streamlit==1.37.1
openai>=1.35.0
python-docx==1.1.2
pypdf==4.3.1
pandas==2.2.2
openpyxl==3.1.5
